Update BookDAO.java
딜리트 메소드 수정

+++ project-DongneBook/src/main/java/model/dao/BookDAO.java
@@ -146,11 +146,9 @@ public class BookDAO {
 	/**
 	 * 북 ID에 해당하는 사용자를 삭제. 
 	 */
-	public int deleteBook(int bookId) throws SQLException {
-		//보류
-		String sql = "DELETE FROM condition WHERE book_id=?; "
-				+ "DELETE FROM book WHERE book_id=?";	
-		jdbcUtil.setSqlAndParameters(sql, new Object[] {bookId});	// JDBCUtil에 delete문과 매개 변수 설정
+	public int deleteBook(int bookId, String userId) throws SQLException {
+		String sql = "DELETE FROM book WHERE book_id=? AND user_id=?";	
+		jdbcUtil.setSqlAndParameters(sql, new Object[] {bookId, userId});	// JDBCUtil에 delete문과 매개 변수 설정
 
 		try {				
 			int result = jdbcUtil.executeUpdate();	// delete 문 실행
